feat: Docker Compose 통합 오케스트레이션 + 빌드 수정 + README 전면 개편
- docker-compose.yaml: TypeScript + Kotlin 동시 기동 (additional_contexts로 외부 의존성 해결)
- typescript/Dockerfile: tomo-idv-client-node를 /deps/로 분리하여 TypeScript 컴파일 간섭 방지
- kotlin/Dockerfile: 3-stage 빌드 (SDK publishToMavenLocal → BFF bootJar → JRE 런타임)
- kotlin/build.gradle.kts: okhttp3, kotlinx-serialization 전이 의존성 추가
- kotlin/TokenService.kt: expiresIn 타입 캐스팅 수정
- typescript/main.ts: TEST_BOARD_PATH 환경변수 지원
- README.md: 멀티 프로젝트 구조, 실행 방법, API 엔드포인트, Docker 빌드 상세 문서화

+++ kotlin/src/main/kotlin/com/tomoarrow/idv/bff/service/TokenService.kt
@@ -45,7 +45,7 @@ class TokenService(
 
         val accessToken = tokenResponse.accessToken ?: ""
         val tokenType = tokenResponse.tokenType ?: "Bearer"
-        val expiresIn = tokenResponse.expiresIn ?: 0L
+        val expiresIn = (tokenResponse.expiresIn ?: 0).toLong()
 
         stateService.set("access_token", accessToken)
         stateService.set("token_info", mapOf(
